refactor: move components, refactor IconMessage
