+++ src/lib/svg/logo/MeLarge.svelte
