+++ src/lib/svg/logo/UniMeLogoLight.svelte
