+++ src/lib/svg/logo/UniMeTextLight.svelte
