+++ src/lib/svg/navbar-icons/Activity.svelte
