+++ src/lib/svg/navbar-icons/Scan.svelte
