+++ src/routes/(app)/activity/connection/[id]/+page@.svelte
@@ -7,8 +7,8 @@
   import ConnectionSummary from '$lib/connections/ConnectionSummary.svelte';
   import type { Connection } from '$lib/connections/types';
   import exampleConnections from '$lib/example/data/connections.json';
+  import ConnectionTabs from '$src/lib/components/molecules/navigation/tabs/ConnectionTabs.svelte';
   import TopNavigation from '$src/lib/components/molecules/navigation/TopNavigation.svelte';
-  import ConnectionTabs from '$src/lib/components/molecules/tabs/ConnectionTabs.svelte';
   import { state } from '$src/stores';
 
   // let connection: Connection = exampleConnections.find((c) => c.id === $page.params.id)!!;
