+++ src/routes/(journey)/goals/+page.svelte
@@ -4,12 +4,12 @@
 
   import { melt } from '@melt-ui/svelte';
 
-  import Button from '$lib/components/Button.svelte';
-  import PaddedIcon from '$lib/components/PaddedIcon.svelte';
   import ProgressBar from '$lib/components/ProgressBar.svelte';
   import { dispatch } from '$lib/dispatcher';
   import GoalItem from '$lib/journey/goals/GoalItem.svelte';
   import LL from '$src/i18n/i18n-svelte';
+  import Button from '$src/lib/components/atoms/Button.svelte';
+  import PaddedIcon from '$src/lib/components/atoms/PaddedIcon.svelte';
   import BottomDrawer from '$src/lib/components/molecules/dialogs/BottomDrawer.svelte';
   import TopNavigation from '$src/lib/components/molecules/navigation/TopNavigation.svelte';
   import { icons } from '$src/lib/journey/goals/utils';
