+++ src/routes/+layout.svelte
@@ -21,7 +21,7 @@
 
   import { melt } from '@melt-ui/svelte';
 
-  import Button from '$src/lib/components/Button.svelte';
+  import Button from '$src/lib/components/atoms/Button.svelte';
   import BottomDrawer from '$src/lib/components/molecules/dialogs/BottomDrawer.svelte';
 
   import { determineTheme } from './utils';
