+++ src/routes/welcome/+layout.svelte
@@ -3,7 +3,7 @@
   import { page } from '$app/stores';
   import { fly } from 'svelte/transition';
 
-  import Button from '$src/lib/components/Button.svelte';
+  import Button from '$src/lib/components/atoms/Button.svelte';
   import TopNavigation from '$src/lib/components/molecules/navigation/TopNavigation.svelte';
 
   import type { LayoutData } from './$types';
