+++ src/routes/welcome/+page.svelte
@@ -3,11 +3,11 @@
   import { fade, fly } from 'svelte/transition';
 
   import LL from '$src/i18n/i18n-svelte';
-  import Button from '$src/lib/components/Button.svelte';
-  import MeLarge from '$src/lib/components/logo/MeLarge.svelte';
-  import UniMeTextDark from '$src/lib/components/logo/UniMeTextDark.svelte';
-  import UniMeTextLight from '$src/lib/components/logo/UniMeTextLight.svelte';
+  import Button from '$src/lib/components/atoms/Button.svelte';
   import { dispatch } from '$src/lib/dispatcher';
+  import MeLarge from '$src/lib/svg/logo/MeLarge.svelte';
+  import UniMeTextDark from '$src/lib/svg/logo/UniMeTextDark.svelte';
+  import UniMeTextLight from '$src/lib/svg/logo/UniMeTextLight.svelte';
   import { onboarding_state, state } from '$src/stores';
 
   // TODO: make reactive
