+++ src/routes/welcome/customize/avatar/+page.svelte
@@ -5,7 +5,7 @@
   import { createPopover, melt } from '@melt-ui/svelte';
 
   import LL from '$src/i18n/i18n-svelte';
-  import Button from '$src/lib/components/Button.svelte';
+  import Button from '$src/lib/components/atoms/Button.svelte';
   import EmojiAvatarSelect from '$src/lib/customize/EmojiAvatarSelect.svelte';
   import { onboarding_state } from '$src/stores';
